new dose presentation

+++ db/seeds.rb
@@ -5,17 +5,18 @@
 #
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
-Ingredient.destroy_all
-Ingredient.create(name: "lemon")
-Ingredient.create(name: "ice")
-Ingredient.create(name: "mint leaves")
-Ingredient.create(name: "apple")
-Ingredient.create(name: "coconut")
-Ingredient.create(name: "sugar")
-Ingredient.create(name: "vodka")
-Ingredient.create(name: "rhum")
-Ingredient.create(name: "tomato")
-Ingredient.create(name: "ginger")
-Ingredient.create(name: "tequila")
-Ingredient.create(name: "water")
+require 'open-uri'
 
+puts "Destroy ingredients"
+Ingredient.destroy_all if Rails.env.development?
+
+puts "Destroy Cocktails"
+Cocktail.destroy_all if Rails.env.development?
+
+puts "Create ingredients"
+url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
+ingredients = JSON.parse(open(url).read)
+ingredients["drinks"].each do |ingredient|
+  i = Ingredient.create(name: ingredient["strIngredient1"])
+  puts "create #{i.name}"
+end
